requirements.txt
Django==2.2.3
celery==4.3.0
djangorestframework==3.10.1
django-configurations==2.1
psycopg2-binary==2.8.3
sentry-sdk==0.10.2
whitenoise==4.1.3
git+ssh://git@bitbucket.com/dataemo/django-origin-trail.git
git+ssh://git@bitbucket.org/dataemo/django_dogstatsd_v2.git
